Convert to module
The information is basically in the README.txt, but basically all the
scripts are in bin, and ssp.py is renamed to core.py.  The module is
actually __init__.py, which imports the right things.  Next step is
split out the different functions of core.py, so the module itself is
more modular.